services/openai_service.py
from openai import AsyncOpenAI
import os
from dotenv import load_dotenv

load_dotenv()
client = AsyncOpenAI(api_key=os.getenv("OpenAI_KEY"))


async def ask_gpt(prompt: str, system_role: str = None, model: str = "gpt-4o-mini", temperature: float = 0.3) -> str:

    messages = []
    if system_role:
        messages.append({"role": "system", "content": system_role})
    messages.append({"role": "user", "content": prompt})

    response = await client.chat.completions.create(
        model=model,
        messages=messages,
        temperature=temperature
    )

    return response.choices[0].message.content.strip()