requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
openai==1.51.2
aiohttp==3.10.5
